requirements.txt
transformers==4.37.2
torch==2.2.0
langchain==0.1.7
langchain-openai==0.0.6
faiss-cpu
tiktoken
python-dotenv